refactor: Update SecurityConfig to suppress deprecation warning, modify JoinChallengeRequest by removing unused constructor, enhance ResponseDto with static factory methods for success and error responses, change Restaurant entity fields from BigDecimal to Double, and adjust AdminServiceImpl and RestaurantServiceImpl to handle new data types and null checks.

+++ src/main/java/com/exe201/color_bites_be/entity/Restaurant.java
@@ -43,10 +43,10 @@ public class Restaurant {
     private String region;
 
     @Field("avg_price")
-    private BigDecimal avgPrice;
+    private Double avgPrice;
 
     @Field("rating")
-    private BigDecimal rating;
+    private Double rating;
 
     @Field("featured")
     @Indexed
